refactor: move bank login configs to co-located files
Each bank's LoginConfig (helpers, consts, factory fn) now lives in a
*LoginConfig.ts file next to its scraper instead of in BankRegistry.ts /
BankRegistryExtra.ts. All loginUrl values source from
SCRAPER_CONFIGURATION.banks[CompanyTypes.X].urls.base — no hardcoded
DNS strings.

- Add 9 new *LoginConfig.ts files (one per bank / bank group)
- Delete BankRegistryExtra.ts
- BankRegistry.ts reduced to a thin 35-line aggregation layer
- All 13 scraper constructors import their own LoginConfig directly
- BaseBeinleumiGroup.test.ts uses beinleumiConfig() directly
