+++ src/Scrapers/Hapoalim/HapoalimScraper.ts
@@ -11,8 +11,8 @@ import { CompanyTypes } from '../../Definitions';
 import { type Transaction, TransactionStatuses, TransactionTypes } from '../../Transactions';
 import { GenericBankScraper } from '../Base/GenericBankScraper';
 import { type ScraperOptions } from '../Base/Interface';
-import { BANK_REGISTRY } from '../Registry/BankRegistry';
 import { SCRAPER_CONFIGURATION } from '../Registry/ScraperConfig';
+import { HAPOALIM_CONFIG } from './HapoalimLoginConfig';
 
 const LOG = getDebug('hapoalim');
 
@@ -303,7 +303,7 @@ interface ScraperSpecificCredentials {
 
 class HapoalimScraper extends GenericBankScraper<ScraperSpecificCredentials> {
   constructor(options: ScraperOptions) {
-    super(options, BANK_REGISTRY[CompanyTypes.Hapoalim]!);
+    super(options, HAPOALIM_CONFIG);
   }
 
   public async fetchData(): Promise<{
