+++ src/Scrapers/Pagi/PagiScraper.ts
@@ -1,11 +1,12 @@
 import { CompanyTypes } from '../../Definitions';
 import { type ScraperOptions } from '../Base/Interface';
 import BeinleumiGroupBaseScraper from '../BaseBeinleumiGroup/BaseBeinleumiGroup';
-import { BANK_REGISTRY } from '../Registry/BankRegistry';
+import { beinleumiConfig } from '../BaseBeinleumiGroup/BeinleumiLoginConfig';
+import { SCRAPER_CONFIGURATION } from '../Registry/ScraperConfig';
 
 class PagiScraper extends BeinleumiGroupBaseScraper {
   constructor(options: ScraperOptions) {
-    super(options, BANK_REGISTRY[CompanyTypes.Pagi]!);
+    super(options, beinleumiConfig(SCRAPER_CONFIGURATION.banks[CompanyTypes.Pagi].urls.base));
   }
 }
