--- src/Scrapers/Registry/BankRegistryExtra.ts
@@ -1,327 +1,0 @@
-import { type Frame, type Page } from 'playwright';
-
-import {
-  clickButton,
-  elementPresentOnPage,
-  fillInput,
-  pageEvalAll,
-  waitUntilElementDisappear,
-  waitUntilElementFound,
-} from '../../Common/ElementsInteractions';
-import { waitForNavigation, waitForRedirect } from '../../Common/Navigation';
-import { resolveFieldContext } from '../../Common/SelectorResolver';
-import { sleep } from '../../Common/Waiting';
-import { CompanyTypes } from '../../Definitions';
-import { type FieldConfig, type LoginConfig } from '../Base/LoginConfig';
-import { SCRAPER_CONFIGURATION } from './ScraperConfig';
-
-const HAPOALIM_BASE = 'https://www.bankhapoalim.co.il';
-
-const LEUMI_INVALID_PASSWORD_MSG = 'אחד או יותר מפרטי ההזדהות שמסרת שגויים. ניתן לנסות שוב';
-const LEUMI_ACCOUNT_BLOCKED_MSG = 'המנוי חסום';
-
-const MIZRAHI_CHECKING_ACCOUNT_HE = 'עובר ושב';
-const MIZRAHI_CHECKING_ACCOUNT_EN = 'Checking Account';
-const MIZRAHI_INVALID_SELECTOR =
-  'a[href*="https://sc.mizrahi-tefahot.co.il/SCServices/SC/P010.aspx"]';
-
-async function leumiCheckReadiness(page: Page): Promise<void> {
-  await waitUntilElementFound(page, '.enter_account');
-  const loginUrl = await page.$eval('.enter_account', el => (el as HTMLAnchorElement).href);
-  await page.goto(loginUrl);
-  await waitForNavigation(page, { waitUntil: 'networkidle' });
-  await Promise.all([
-    waitUntilElementFound(page, 'input[placeholder="שם משתמש"]', { visible: true }),
-    waitUntilElementFound(page, 'input[placeholder="סיסמה"]', { visible: true }),
-    waitUntilElementFound(page, 'button[type="submit"]', { visible: true }),
-  ]);
-}
-
-async function leumiPostAction(page: Page): Promise<void> {
-  await Promise.race([
-    waitUntilElementFound(page, 'a[title="דלג לחשבון"]', { visible: true, timeout: 60000 }),
-    waitUntilElementFound(page, 'div.main-content', { visible: false, timeout: 60000 }),
-    page.waitForSelector(`xpath=//div[contains(string(),"${LEUMI_INVALID_PASSWORD_MSG}")]`),
-    waitUntilElementFound(page, 'form[action="/changepassword"]', {
-      visible: true,
-      timeout: 60000,
-    }),
-  ]);
-}
-
-async function mizrahiIsLoggedIn(opts?: { page?: Page }): Promise<boolean> {
-  if (!opts?.page) return false;
-  const xpath = `//a//span[contains(., "${MIZRAHI_CHECKING_ACCOUNT_HE}") or contains(., "${MIZRAHI_CHECKING_ACCOUNT_EN}")]`;
-  return (await opts.page.$$(`xpath=${xpath}`)).length > 0;
-}
-
-async function mizrahiPostAction(page: Page): Promise<void> {
-  await Promise.race([
-    waitUntilElementFound(page, '#dropdownBasic'),
-    waitUntilElementFound(page, MIZRAHI_INVALID_SELECTOR),
-    waitForNavigation(page),
-  ]);
-}
-
-async function maxPreActionStep1(page: Page): Promise<void> {
-  if (await elementPresentOnPage(page, '#closePopup'))
-    await page.$eval('#closePopup', el => {
-      (el as HTMLElement).click();
-    });
-  await page.$eval('.personal-area > a.go-to-personal-area', el => {
-    (el as HTMLElement).click();
-  });
-}
-
-async function maxPreActionStep2(page: Page): Promise<void> {
-  if (await elementPresentOnPage(page, '.login-link#private'))
-    await page.$eval('.login-link#private', el => {
-      (el as HTMLElement).click();
-    });
-  await waitUntilElementFound(page, '#login-password-link', { visible: true });
-  await page.$eval('#login-password-link', el => {
-    (el as HTMLElement).click();
-  });
-  await waitUntilElementFound(page, '#login-password.tab-pane.active app-user-login-form', {
-    visible: true,
-  });
-}
-
-async function maxPreAction(page: Page): Promise<Frame | undefined> {
-  await maxPreActionStep1(page);
-  await maxPreActionStep2(page);
-  return undefined;
-}
-
-// FieldConfig for the ID field — selectors empty so SelectorResolver falls back to wellKnownSelectors.id
-const MAX_ID_FIELD_CONFIG: FieldConfig = { credentialKey: 'id', selectors: [] };
-
-/**
- * Handles the optional second-login step in Max's Flow B:
- *   home → username+password → 2nd form (username+password+ID) → dashboard
- * Uses wellKnownSelectors.id from ScraperConfig to detect the field.
- * If the ID form is not present (Flow A), this function is a no-op.
- */
-export async function maxHandleSecondLoginStep(
-  page: Page,
-  credentials: { username: string; password: string; id?: string },
-): Promise<void> {
-  if (!credentials.id) return;
-  const idCtx = await resolveFieldContext(page, MAX_ID_FIELD_CONFIG, page.url());
-  if (!idCtx.isResolved) return;
-  await fillInput(page, '#user-name', credentials.username);
-  await fillInput(page, '#password', credentials.password);
-  await fillInput(idCtx.context, idCtx.selector, credentials.id);
-  await clickButton(page, 'app-user-login-form .general-button.send-me-code');
-  await sleep(1000);
-}
-
-async function maxPostAction(page: Page): Promise<void> {
-  if (page.url().includes('/homepage/personal')) return;
-  await Promise.race([
-    waitForRedirect(page, {
-      timeout: 20000,
-      ignoreList: ['https://www.max.co.il', 'https://www.max.co.il/'],
-    }),
-    waitUntilElementFound(page, '#popupWrongDetails', { visible: true }),
-    waitUntilElementFound(page, '#popupCardHoldersLoginError', { visible: true }),
-  ]);
-}
-
-async function yahavPostAction(page: Page): Promise<void> {
-  await waitForNavigation(page);
-  await waitUntilElementDisappear(page, '.loader');
-  if (await elementPresentOnPage(page, '.messaging-links-container')) {
-    await page.$eval('.link-1', el => {
-      (el as HTMLElement).click();
-    });
-  }
-  await Promise.race([
-    waitUntilElementFound(page, '#AccountDetails'),
-    waitUntilElementFound(page, 'input#ef_req_parameter_old_credential'),
-  ]);
-}
-
-export const HAPOALIM_CONFIG: LoginConfig = {
-  loginUrl: HAPOALIM_BASE,
-  fields: [
-    { credentialKey: 'userCode', selectors: [] }, // wellKnown → #userCode
-    { credentialKey: 'password', selectors: [] }, // wellKnown → #password
-  ],
-  submit: [
-    { kind: 'css', value: '.login-btn' },
-    // ariaLabel 'כניסה' fallback is now in wellKnownSelectors.__submit__
-  ],
-  postAction: async (page: Page) => {
-    await waitForRedirect(page, {});
-  },
-  possibleResults: {
-    success: [
-      'https://login.bankhapoalim.co.il/portalserver/HomePage',
-      'https://login.bankhapoalim.co.il/ng-portals-bt/rb/he/homepage',
-      'https://login.bankhapoalim.co.il/ng-portals/rb/he/homepage',
-    ],
-    invalidPassword: [
-      'https://login.bankhapoalim.co.il/AUTHENTICATE/LOGON?flow=AUTHENTICATE&state=LOGON&errorcode=1.6&callme=false',
-    ],
-    changePassword: [
-      'https://login.bankhapoalim.co.il/MCP/START?flow=MCP&state=START&expiredDate=null',
-      /\/ABOUTTOEXPIRE\/START/i,
-    ],
-  },
-};
-
-export const LEUMI_CONFIG: LoginConfig = {
-  loginUrl: 'https://www.leumi.co.il',
-  fields: [
-    { credentialKey: 'username', selectors: [] }, // wellKnown → placeholder שם משתמש
-    { credentialKey: 'password', selectors: [] }, // wellKnown → placeholder סיסמה
-  ],
-  submit: [{ kind: 'css', value: "button[type='submit']" }],
-  checkReadiness: leumiCheckReadiness,
-  postAction: leumiPostAction,
-  possibleResults: {
-    success: [/ebanking\/SO\/SPA.aspx/i],
-    invalidPassword: [
-      async (opts): Promise<boolean> => {
-        if (!opts?.page) return false;
-        const msg = await pageEvalAll(opts.page, {
-          selector: 'svg#Capa_1',
-          defaultResult: '',
-          callback: el => (el[0]?.parentElement?.children[1] as HTMLDivElement).innerText,
-        });
-        return msg.startsWith(LEUMI_INVALID_PASSWORD_MSG);
-      },
-    ],
-    accountBlocked: [
-      async (opts): Promise<boolean> => {
-        if (!opts?.page) return false;
-        const msg = await pageEvalAll(opts.page, {
-          selector: '.errHeader',
-          defaultResult: '',
-          callback: el => (el[0] as HTMLElement).innerText,
-        });
-        return msg.startsWith(LEUMI_ACCOUNT_BLOCKED_MSG);
-      },
-    ],
-    changePassword: ['https://hb2.bankleumi.co.il/authenticate'],
-  },
-};
-
-const MIZRAHI_LOGIN_ROUTE = SCRAPER_CONFIGURATION.banks[CompanyTypes.Mizrahi].urls.loginRoute;
-
-export const MIZRAHI_CONFIG: LoginConfig = {
-  loginUrl: 'https://www.mizrahi-tefahot.co.il',
-  fields: [
-    { credentialKey: 'username', selectors: [{ kind: 'css', value: '#userNumberDesktopHeb' }] },
-    { credentialKey: 'password', selectors: [{ kind: 'css', value: '#passwordDesktopHeb' }] },
-  ],
-  submit: [{ kind: 'css', value: 'button.btn.btn-primary' }],
-  checkReadiness: async (page: Page) => {
-    await page.goto(MIZRAHI_LOGIN_ROUTE, { waitUntil: 'domcontentloaded' });
-    await waitUntilElementDisappear(page, 'div.ngx-overlay.loading-foreground');
-  },
-  postAction: mizrahiPostAction,
-  possibleResults: {
-    success: [/https:\/\/mto\.mizrahi-tefahot\.co\.il\/OnlineApp\/.*/i, mizrahiIsLoggedIn],
-    invalidPassword: [
-      async (opts): Promise<boolean> =>
-        !!(opts?.page && (await opts.page.$(MIZRAHI_INVALID_SELECTOR))),
-    ],
-    changePassword: [/https:\/\/www\.mizrahi-tefahot\.co\.il\/login\/index\.html#\/change-pass/],
-  },
-};
-
-export const MAX_CONFIG: LoginConfig = {
-  loginUrl: 'https://www.max.co.il',
-  fields: [
-    { credentialKey: 'username', selectors: [] }, // wellKnown → #user-name
-    { credentialKey: 'password', selectors: [] }, // wellKnown → #password
-  ],
-  submit: [{ kind: 'css', value: 'app-user-login-form .general-button.send-me-code' }],
-  checkReadiness: async (page: Page) => {
-    await waitUntilElementFound(page, '.personal-area > a.go-to-personal-area', { visible: true });
-  },
-  preAction: maxPreAction,
-  postAction: maxPostAction,
-  waitUntil: 'domcontentloaded',
-  possibleResults: {
-    success: ['https://www.max.co.il/homepage/personal'],
-    changePassword: ['https://www.max.co.il/renew-password'],
-    invalidPassword: [
-      async (opts): Promise<boolean> =>
-        !!(opts?.page && (await elementPresentOnPage(opts.page, '#popupWrongDetails'))),
-    ],
-    unknownError: [
-      async (opts): Promise<boolean> =>
-        !!(opts?.page && (await elementPresentOnPage(opts.page, '#popupCardHoldersLoginError'))),
-    ],
-  },
-};
-
-// Behatsdaa and BeyahadBishvilha share the same login form selectors
-// (selectors: [] — wellKnown finds #loginId and #loginPassword)
-const HISTBASED_FIELDS: LoginConfig['fields'] = [
-  { credentialKey: 'id', selectors: [] }, // wellKnown → #loginId
-  { credentialKey: 'password', selectors: [] }, // wellKnown → #loginPassword
-];
-
-export const BEHATSDAA_CONFIG: LoginConfig = {
-  loginUrl: 'https://www.behatsdaa.org.il',
-  fields: HISTBASED_FIELDS,
-  submit: [
-    { kind: 'xpath', value: '//button[contains(., "התחברות")]' },
-    // ariaLabel 'התחברות' fallback is now in wellKnownSelectors.__submit__
-  ],
-  checkReadiness: async (page: Page) => {
-    await page.goto('https://www.behatsdaa.org.il/login');
-  },
-  possibleResults: {
-    success: ['https://www.behatsdaa.org.il/'],
-    invalidPassword: [
-      async (opts): Promise<boolean> =>
-        !!(opts?.page && (await elementPresentOnPage(opts.page, '.custom-input-error-label'))),
-    ],
-  },
-};
-
-export const BEYAHAD_CONFIG: LoginConfig = {
-  loginUrl: 'https://www.hist.org.il',
-  fields: HISTBASED_FIELDS,
-  submit: [
-    { kind: 'xpath', value: '//button[contains(., "התחבר")]' },
-    // ariaLabel 'התחבר' fallback is now in wellKnownSelectors.__submit__
-  ],
-  checkReadiness: async (page: Page) => {
-    await page.goto('https://www.hist.org.il/login');
-  },
-  possibleResults: { success: ['https://www.hist.org.il/'] },
-};
-
-export const YAHAV_CONFIG: LoginConfig = {
-  loginUrl: 'https://www.yahav.co.il',
-  fields: [
-    { credentialKey: 'username', selectors: [] }, // wellKnown → #username
-    { credentialKey: 'password', selectors: [] }, // wellKnown → #password
-    { credentialKey: 'nationalID', selectors: [] }, // wellKnown → #pinno
-  ],
-  submit: [{ kind: 'css', value: '.btn' }],
-  checkReadiness: async (page: Page) => {
-    await Promise.all([waitUntilElementFound(page, '#pinno'), waitUntilElementFound(page, '.btn')]);
-  },
-  postAction: yahavPostAction,
-  possibleResults: {
-    success: ['https://digital.yahav.co.il/BaNCSDigitalUI/app/index.html#/main/home'],
-    invalidPassword: [
-      async (opts): Promise<boolean> =>
-        !!(opts?.page && (await elementPresentOnPage(opts.page, '.ui-dialog-buttons'))),
-    ],
-    changePassword: [
-      async (opts): Promise<boolean> =>
-        !!(
-          opts?.page &&
-          (await elementPresentOnPage(opts.page, 'input#ef_req_parameter_old_credential'))
-        ),
-    ],
-  },
-};
