+++ src/Scrapers/Yahav/YahavScraper.ts
@@ -19,8 +19,8 @@ import {
 } from '../../Transactions';
 import { GenericBankScraper } from '../Base/GenericBankScraper';
 import { type ScraperOptions } from '../Base/Interface';
-import { BANK_REGISTRY } from '../Registry/BankRegistry';
 import { SCRAPER_CONFIGURATION } from '../Registry/ScraperConfig';
+import { YAHAV_CONFIG } from './YahavLoginConfig';
 
 const CFG = SCRAPER_CONFIGURATION.banks[CompanyTypes.Yahav];
 // Phase-1 compat: extract first CSS candidate until full resolveDashboardField() migration
@@ -223,7 +223,7 @@ interface ScraperSpecificCredentials {
 
 class YahavScraper extends GenericBankScraper<ScraperSpecificCredentials> {
   constructor(options: ScraperOptions) {
-    super(options, BANK_REGISTRY[CompanyTypes.Yahav]!);
+    super(options, YAHAV_CONFIG);
   }
 
   public async fetchData(): Promise<{ success: boolean; accounts: TransactionsAccount[] }> {
